Implement get_job_plans()
